[WIP] Update rad init --dev experience
