+++ pkg/corerp/frontend/controller/environments/types.go
@@ -6,16 +6,7 @@
 package environments
 
 const (
-	ResourceTypeName  = "Applications.Core/environments"
-	DevRecipesACRPath = "radius.azurecr.io"
+	ResourceTypeName = "Applications.Core/environments"
 	// User defined operation names
 	OperationGetRecipeMetadata = "GETRECIPEMETADATA"
 )
-
-// supportedProviders returns the list of "known" providers we understand for dev recipes.
-// this is used as a filter to exclude non-matching repositories from the dev recipes registry.
-//
-// This is no effect on the execution of the recipe.
-func supportedProviders() []string {
-	return []string{"aws", "azure", "kubernetes"}
-}
